+++ examples/simple/expected/MyScalarQuery.php
@@ -8,9 +8,7 @@ class MyScalarQuery extends \Spawnia\Sailor\Operation
 {
     public static function execute(?string $arg = null): MyScalarQuery\MyScalarQueryResult
     {
-        $response = self::fetchResponse(...func_get_args());
-
-        return \Spawnia\Sailor\Simple\MyScalarQuery\MyScalarQueryResult::fromResponse($response);
+        return self::executeOperation(...func_get_args());
     }
 
     public static function document(): string