+++ tests/Unit/ResultTest.php
@@ -6,6 +6,7 @@ namespace Spawnia\Sailor\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 use Spawnia\Sailor\ResultErrorsException;
+use Spawnia\Sailor\Simple\MyScalarQuery\MyScalarQuery;
 use Spawnia\Sailor\Simple\MyScalarQuery\MyScalarQueryResult;
 
 class ResultTest extends TestCase
@@ -21,5 +22,21 @@ class ResultTest extends TestCase
 
         $this->expectException(ResultErrorsException::class);
         $result->assertErrorFree();
+    }
+
+    public function testErrorFree(): void
+    {
+        $result = new MyScalarQueryResult();
+        $result->data = MyScalarQuery::fromStdClass((object) [
+            'scalarWithArg' => null,
+        ]);
+
+        // No errors
+        $result->errorFree();
+
+        $result->errors = [new \stdClass()];
+
+        $this->expectException(ResultErrorsException::class);
+        $result->errorFree();
     }
 }