+++ CHANGELOG.md
@@ -6,10 +6,31 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## v0.4.0
+
+### Added
+
+- Add method `Result::errorFree()` to ease safe access to a result class without errors
+
+## v0.3.0
+
+### Added
+
+- Allow mocking single operations with Mockery
+
+### Fixed
+
+- Fix codegen for operations with multiple parameters
+
 ### Changed
 
 - Require `Directive.isRepeatable` on introspection
 - Use typed properties
+- Rewrite `Configuration` class
+
+### Removed
+
+- Remove `EndpointConfig::$mockClient` and `EndpointConfig::client()`
 
 ## v0.2.0